+++ common/src/main/java/dev/xylonity/bonsai/clockwork/client/projectile/model/GenericProjectileModel.java
@@ -1,24 +1,24 @@
 package dev.xylonity.bonsai.clockwork.client.projectile.model;
 
 import dev.xylonity.bonsai.clockwork.Clockwork;
-import dev.xylonity.knightlib.common.entity.AbstractProjectile;
+import dev.xylonity.bonsai.clockwork.common.entity.projectile.GenericProjectile;
 import net.minecraft.resources.ResourceLocation;
 import software.bernie.geckolib.model.GeoModel;
 
-public class GenericProjectileModel<T extends AbstractProjectile> extends GeoModel<T> {
+public class GenericProjectileModel<T extends GenericProjectile> extends GeoModel<T> {
 
     @Override
-    public ResourceLocation getModelResource(AbstractProjectile baseProjectile) {
+    public ResourceLocation getModelResource(GenericProjectile baseProjectile) {
         return Clockwork.resource("geo/generic.geo.json");
     }
 
     @Override
-    public ResourceLocation getTextureResource(AbstractProjectile baseProjectile) {
+    public ResourceLocation getTextureResource(GenericProjectile baseProjectile) {
         return Clockwork.resource("textures/entity/generic.png");
     }
 
     @Override
-    public ResourceLocation getAnimationResource(AbstractProjectile baseProjectile) {
+    public ResourceLocation getAnimationResource(GenericProjectile baseProjectile) {
         return Clockwork.resource("animations/generic.animation.json");
     }