+++ common/src/main/java/dev/xylonity/bonsai/clockwork/common/entity/projectile/GenericProjectile.java
@@ -1,0 +1,30 @@
+package dev.xylonity.bonsai.clockwork.common.entity.projectile;
+
+import dev.xylonity.knightlib.common.entity.AbstractProjectile;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.projectile.Projectile;
+import net.minecraft.world.level.Level;
+import software.bernie.geckolib.animatable.GeoEntity;
+import software.bernie.geckolib.animatable.instance.AnimatableInstanceCache;
+import software.bernie.geckolib.animation.AnimatableManager;
+import software.bernie.geckolib.util.GeckoLibUtil;
+
+public abstract class GenericProjectile extends AbstractProjectile implements GeoEntity {
+
+    private final AnimatableInstanceCache cache = GeckoLibUtil.createInstanceCache(this);
+
+    public GenericProjectile(EntityType<? extends Projectile> pEntityType, Level pLevel) {
+        super(pEntityType, pLevel);
+    }
+
+    @Override
+    public AnimatableInstanceCache getAnimatableInstanceCache() {
+        return cache;
+    }
+
+    @Override
+    public void registerControllers(AnimatableManager.ControllerRegistrar controllers) {
+        ;;
+    }
+
+}