+++ common/src/main/java/dev/xylonity/bonsai/clockwork/common/entity/projectile/trigger/GenericTriggerProjectile.java
@@ -1,6 +1,6 @@
 package dev.xylonity.bonsai.clockwork.common.entity.projectile.trigger;
 
-import dev.xylonity.knightlib.common.entity.AbstractProjectile;
+import dev.xylonity.bonsai.clockwork.common.entity.projectile.GenericProjectile;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.player.Player;
@@ -9,7 +9,7 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.NotNull;
 
-public class GenericTriggerProjectile extends AbstractProjectile {
+public class GenericTriggerProjectile extends GenericProjectile {
 
     public GenericTriggerProjectile(EntityType<? extends Projectile> pEntityType, Level pLevel) {
         super(pEntityType, pLevel);